moved locallib test files to a single file
