+++ tests/locallib_test.php
@@ -46,10 +46,11 @@ class locallib_test extends \advanced_testcase {
             'name' => 'new_digitala',
             'attemptlang' => 'fin',
             'attempttype' => 'freeform',
-            'assignment' => array('text' => 'Assignment text', 'format' => 1),
+            'assignment' => 'Assignment text',
             'resources' => array('text' => 'Resource text', 'format' => 1),
         ]);
     }
+
     /**
      * Test generating page_url
      */
@@ -64,19 +65,21 @@ class locallib_test extends \advanced_testcase {
      * Test generating progress bar step link as non active and active
      */
     public function test_create_progress_bar_step_link() {
-        $info = create_progress_bar_step_link('digitalainfo', 0, 1, 1, false);
-        $assignment = create_progress_bar_step_link('digitalaassignment', 1, 1, 1, false);
-        $report = create_progress_bar_step_link('digitalareport', 2, 1, 1, false);
-        $this->assertEquals($info, '<a class="display-6" href="https://www.example.com/moodle/mod/digitala/view.php?id=1&amp;d=1&amp;page=0"><span class="pb-num">1</span>Info</a>'); // phpcs:ignore moodle.Files.LineLength.MaxExceeded
-        $this->assertEquals($assignment, '<a class="display-6" href="https://www.example.com/moodle/mod/digitala/view.php?id=1&amp;d=1&amp;page=1"><span class="pb-num">2</span>Assignment</a>'); // phpcs:ignore moodle.Files.LineLength.MaxExceeded
-        $this->assertEquals($report, '<a class="display-6" href="https://www.example.com/moodle/mod/digitala/view.php?id=1&amp;d=1&amp;page=2"><span class="pb-num">3</span>Report</a>'); // phpcs:ignore moodle.Files.LineLength.MaxExceeded
+        $info = create_progress_bar_step_link('info', 0, 1, 1, false);
+        $assignment = create_progress_bar_step_link('assignment', 1, 1, 1, false);
+        $report = create_progress_bar_step_link('report', 2, 1, 1, false);
+        // @codingStandardsIgnoreStart moodle.Files.LineLength.MaxExceeded
+        $this->assertEquals($info, '<a class="display-6" href="https://www.example.com/moodle/mod/digitala/view.php?id=1&amp;d=1&amp;page=0"><span class="pb-num">1</span>Info</a>');
+        $this->assertEquals($assignment, '<a class="display-6" href="https://www.example.com/moodle/mod/digitala/view.php?id=1&amp;d=1&amp;page=1"><span class="pb-num">2</span>Assignment</a>');
+        $this->assertEquals($report, '<a class="display-6" href="https://www.example.com/moodle/mod/digitala/view.php?id=1&amp;d=1&amp;page=2"><span class="pb-num">3</span>Report</a>');
 
-        $infoactive = create_progress_bar_step_link('digitalainfo', 0, 1, 1, true);
-        $assignmentactive = create_progress_bar_step_link('digitalaassignment', 1, 1, 1, true);
-        $reportactive = create_progress_bar_step_link('digitalareport', 2, 1, 1, true);
-        $this->assertEquals($infoactive, '<a class="display-6" href="https://www.example.com/moodle/mod/digitala/view.php?id=1&amp;d=1&amp;page=0"><span class="pb-num active">1</span>Info</a>'); // phpcs:ignore moodle.Files.LineLength.MaxExceeded
-        $this->assertEquals($assignmentactive, '<a class="display-6" href="https://www.example.com/moodle/mod/digitala/view.php?id=1&amp;d=1&amp;page=1"><span class="pb-num active">2</span>Assignment</a>'); // phpcs:ignore moodle.Files.LineLength.MaxExceeded
-        $this->assertEquals($reportactive, '<a class="display-6" href="https://www.example.com/moodle/mod/digitala/view.php?id=1&amp;d=1&amp;page=2"><span class="pb-num active">3</span>Report</a>'); // phpcs:ignore moodle.Files.LineLength.MaxExceeded
+        $infoactive = create_progress_bar_step_link('info', 0, 1, 1, true);
+        $assignmentactive = create_progress_bar_step_link('assignment', 1, 1, 1, true);
+        $reportactive = create_progress_bar_step_link('report', 2, 1, 1, true);
+        $this->assertEquals($infoactive, '<a class="display-6" href="https://www.example.com/moodle/mod/digitala/view.php?id=1&amp;d=1&amp;page=0"><span class="pb-num active">1</span>Info</a>');
+        $this->assertEquals($assignmentactive, '<a class="display-6" href="https://www.example.com/moodle/mod/digitala/view.php?id=1&amp;d=1&amp;page=1"><span class="pb-num active">2</span>Assignment</a>');
+        $this->assertEquals($reportactive, '<a class="display-6" href="https://www.example.com/moodle/mod/digitala/view.php?id=1&amp;d=1&amp;page=2"><span class="pb-num active">3</span>Report</a>');
+        // @codingStandardsIgnoreEnd moodle.Files.LineLength.MaxExceeded
     }
 
     /**
@@ -99,22 +102,24 @@ class locallib_test extends \advanced_testcase {
      * Test generating whole step in the progress bar
      */
     public function test_create_progress_bar_step() {
-        $info = create_progress_bar_step('digitalainfo', 0, 1, 1, 1);
-        $assignment = create_progress_bar_step('digitalaassignment', 1, 1, 1, 0);
-        $report = create_progress_bar_step('digitalareport', 2, 1, 1, 0);
-        $infoactive = create_progress_bar_step('digitalainfo', 0, 1, 1, 0);
-        $assignmentactive = create_progress_bar_step('digitalaassignment', 1, 1, 1, 1);
-        $reportactive = create_progress_bar_step('digitalareport', 2, 1, 1, 2);
-        $this->assertEquals($info, '<div class="pb-step first"><a class="display-6" href="https://www.example.com/moodle/mod/digitala/view.php?id=1&amp;d=1&amp;page=0"><span class="pb-num">1</span>Info</a></div>'); // phpcs:ignore moodle.Files.LineLength.MaxExceeded
-        $this->assertEquals($assignment, '<div class="pb-step"><a class="display-6" href="https://www.example.com/moodle/mod/digitala/view.php?id=1&amp;d=1&amp;page=1"><span class="pb-num">2</span>Assignment</a></div>'); // phpcs:ignore moodle.Files.LineLength.MaxExceeded
-        $this->assertEquals($report, '<div class="pb-step last"><a class="display-6" href="https://www.example.com/moodle/mod/digitala/view.php?id=1&amp;d=1&amp;page=2"><span class="pb-num">3</span>Report</a></div>'); // phpcs:ignore moodle.Files.LineLength.MaxExceeded
-        $this->assertEquals($infoactive, '<div class="pb-step active first"><a class="display-6" href="https://www.example.com/moodle/mod/digitala/view.php?id=1&amp;d=1&amp;page=0"><span class="pb-num active">1</span>Info</a></div>'); // phpcs:ignore moodle.Files.LineLength.MaxExceeded
-        $this->assertEquals($assignmentactive, '<div class="pb-step active"><a class="display-6" href="https://www.example.com/moodle/mod/digitala/view.php?id=1&amp;d=1&amp;page=1"><span class="pb-num active">2</span>Assignment</a></div>'); // phpcs:ignore moodle.Files.LineLength.MaxExceeded
-        $this->assertEquals($reportactive, '<div class="pb-step active last"><a class="display-6" href="https://www.example.com/moodle/mod/digitala/view.php?id=1&amp;d=1&amp;page=2"><span class="pb-num active">3</span>Report</a></div>'); // phpcs:ignore moodle.Files.LineLength.MaxExceeded
+        $info = create_progress_bar_step('info', 0, 1, 1, 1);
+        $assignment = create_progress_bar_step('assignment', 1, 1, 1, 0);
+        $report = create_progress_bar_step('report', 2, 1, 1, 0);
+        $infoactive = create_progress_bar_step('info', 0, 1, 1, 0);
+        $assignmentactive = create_progress_bar_step('assignment', 1, 1, 1, 1);
+        $reportactive = create_progress_bar_step('report', 2, 1, 1, 2);
+        // @codingStandardsIgnoreStart moodle.Files.LineLength.MaxExceeded
+        $this->assertEquals($info, '<div class="pb-step first"><a class="display-6" href="https://www.example.com/moodle/mod/digitala/view.php?id=1&amp;d=1&amp;page=0"><span class="pb-num">1</span>Info</a></div>');
+        $this->assertEquals($assignment, '<div class="pb-step"><a class="display-6" href="https://www.example.com/moodle/mod/digitala/view.php?id=1&amp;d=1&amp;page=1"><span class="pb-num">2</span>Assignment</a></div>');
+        $this->assertEquals($report, '<div class="pb-step last"><a class="display-6" href="https://www.example.com/moodle/mod/digitala/view.php?id=1&amp;d=1&amp;page=2"><span class="pb-num">3</span>Report</a></div>');
+        $this->assertEquals($infoactive, '<div class="pb-step active first"><a class="display-6" href="https://www.example.com/moodle/mod/digitala/view.php?id=1&amp;d=1&amp;page=0"><span class="pb-num active">1</span>Info</a></div>');
+        $this->assertEquals($assignmentactive, '<div class="pb-step active"><a class="display-6" href="https://www.example.com/moodle/mod/digitala/view.php?id=1&amp;d=1&amp;page=1"><span class="pb-num active">2</span>Assignment</a></div>');
+        $this->assertEquals($reportactive, '<div class="pb-step active last"><a class="display-6" href="https://www.example.com/moodle/mod/digitala/view.php?id=1&amp;d=1&amp;page=2"><span class="pb-num active">3</span>Report</a></div>');
+        // @codingStandardsIgnoreEnd moodle.Files.LineLength.MaxExceeded
     }
 
     /**
-     * Test generating page_url
+     * Test calculating progress bar spacers
      */
     public function test_calculate_progress_bar_spacers() {
         $info = calculate_progress_bar_spacers(0);
@@ -126,7 +131,7 @@ class locallib_test extends \advanced_testcase {
     }
 
     /**
-     * Test generating page_url
+     * Test creating progressbar spacer
      */
     public function test_create_progress_bar_spacer() {
         $rightempty = create_progress_bar_spacer('right-empty');
@@ -139,6 +144,26 @@ class locallib_test extends \advanced_testcase {
     style="fill-rule:evenodd;clip-rule:evenodd;stroke-linecap:round;stroke-linejoin:round;stroke-miterlimit:1.5;"><path d="M275,0L20,0L255,250L20,500L275,500L275,0Z" style="fill:rgb(211,211,211);"/><path d="M20,20L255,250L20,480" style="fill:none;stroke:rgb(211,211,211);stroke-width:40px;"/></svg></div>');
         $this->assertEquals($nothing, '<div class="pb-spacer"><svg width="100%" height="100%" viewBox="0 0 275 500"
     style="fill-rule:evenodd;clip-rule:evenodd;stroke-linecap:round;stroke-linejoin:round;stroke-miterlimit:1.5;"><path d="M20,20L255,250L20,480" style="fill:none;stroke:rgb(211,211,211);stroke-width:40px;"/></svg></div>');
+        // @codingStandardsIgnoreEnd moodle.Files.LineLength.MaxExceeded
+    }
+
+    /**
+     * Test creating navigation buttons for view
+     */
+    public function test_navbuttons_html_output() {
+        $result = create_nav_buttons('info', 1, 2);
+        // @codingStandardsIgnoreStart moodle.Files.LineLength.MaxExceeded
+        $this->assertEquals('<div class="navbuttons"><a href=https://www.example.com/moodle/mod/digitala/view.php?id=1&amp;d=2&amp;page=1 id="nextButton" class="btn btn-primary">Next ></a href=https://www.example.com/moodle/mod/digitala/view.php?id=1&amp;d=2&amp;page=1></div>',
+            $result);
+        $result = create_nav_buttons('assignmentprev', 1, 2);
+        $this->assertEquals('<div class="navbuttons"><a href=https://www.example.com/moodle/mod/digitala/view.php?id=1&amp;d=2&amp;page=0 id="prevButton" class="btn btn-primary">< Previous</a href=https://www.example.com/moodle/mod/digitala/view.php?id=1&amp;d=2&amp;page=0></div>',
+            $result);
+        $result = create_nav_buttons('assignmentnext', 1, 2);
+        $this->assertEquals('<div class="navbuttons"><a href=https://www.example.com/moodle/mod/digitala/view.php?id=1&amp;d=2&amp;page=2 id="nextButton" class="btn btn-primary">Next ></a href=https://www.example.com/moodle/mod/digitala/view.php?id=1&amp;d=2&amp;page=2></div>',
+            $result);
+        $result = create_nav_buttons('report', 1, 2);
+        $this->assertEquals('<div class="navbuttons"><a href=https://www.example.com/moodle/mod/digitala/view.php?id=1&amp;d=2&amp;page=1 id="tryAgainButton" class="btn btn-primary">See the assignment</a href=https://www.example.com/moodle/mod/digitala/view.php?id=1&amp;d=2&amp;page=1><a href=https://link.webropolsurveys.com/Participation/Public/2c1ccd52-6e23-436e-af51-f8f8c259ffbb?displayId=Fin2500048 id="feedbackButton" class="btn btn-primary" target="blank">Give feedback</a href=https://link.webropolsurveys.com/Participation/Public/2c1ccd52-6e23-436e-af51-f8f8c259ffbb?displayId=Fin2500048></div>',
+            $result);
         // @codingStandardsIgnoreEnd moodle.Files.LineLength.MaxExceeded
     }
 
@@ -180,13 +205,46 @@ class locallib_test extends \advanced_testcase {
     }
 
     /**
-     * Test creating report view specific helper object.
+     * Test creating report view grading helper object.
      */
     public function test_grading_html_output() {
-        $result = create_report_grading("Grading", 0, 0);
-        $this->assertEquals('<div class="card row digitala-card"><div class="card-body"><h5 class="card-title">Grading</h5>'.
+        $result = create_report_grading("fluency", 0, 0);
+        $this->assertEquals('<div class="card row digitala-card"><div class="card-body"><h5 class="card-title">Fluency</h5>'.
             '<h5 class="grade-stars"></h5><h6 class="grade-number">0/0</h6>'.
-            '<div class="card-text">Grading information will be shown here once they\'re available.</div></div></div>', $result);
+            '<div class="card-text">Fluency score is 0, red score.</div></div></div>', $result);
+    }
+
+    /**
+     * Test creating report view holistic helper object.
+     */
+    public function test_holistic_html_output() {
+        $result = create_report_holistic(3);
+        $this->assertEquals('<div class="card row digitala-card"><div class="card-body"><h5 class="card-title">Holistic</h5>'.
+            '<h6 class="grade-number">B1</h6><div class="card-text">Holistic score is 3, yellow score.</div></div></div>', $result);
+    }
+
+    /**
+     * Test creating report view GOP helper object.
+     */
+    public function test_gop_html_output() {
+        $result = create_report_gop(0.72);
+        $this->assertEquals('<div class="card row digitala-card"><div class="card-body"><h5 class="card-title">Goodness of pronunciation</h5><h6 class="grade-number">72/100</h6><div class="card-text">Pronunciation score is 7, big pink score.</div></div></div>', $result); // phpcs:ignore moodle.Files.LineLength.MaxExceeded
+    }
+
+    /**
+     * Test creating report view tab creation helper.
+     */
+    public function test_tabs_html_output() {
+        $result = create_report_tabs('', '');
+        $this->assertEquals('<nav><div class="nav nav-tabs" id="nav-tab" role="tablist">'.
+            '<button class="nav-link active ml-2" id="report-grades-tab" data-toggle="tab" href="#report-grades" role="tab" '.
+            'aria-controls="report-grades" aria-selected="true">Task grades</button>'.
+            '<button class="nav-link ml-2" id="report-holistic-tab" data-toggle="tab" href="#report-holistic" role="tab" '.
+            'aria-controls="report-holistic" aria-selected="false">Holistic</button>'.
+            '</div></nav><div class="tab-content" id="nav-tabContent">'.
+            '<div class="tab-pane fade show active" id="report-grades" role="tabpanel" aria-labelledby="report-grades-tab"></div>'.
+            '<div class="tab-pane fade" id="report-holistic" role="tabpanel" aria-labelledby="report-holistic-tab">'.
+            '</div></div>', $result);
     }
 
     /**
@@ -200,19 +258,18 @@ class locallib_test extends \advanced_testcase {
     }
 
     /**
-     * Test creating new button.
+     * Test creating new assignment object.
      */
     public function test_create_button() {
-        $result = create_button('buttonId', 'buttonText');
-        $this->assertEquals('<button id="buttonId" class="buttonId">buttonText</button>', $result);
+        $result = create_button('buttonId', 'buttonClass', 'buttonText');
+        $this->assertEquals('<button id="buttonId" class="buttonClass">buttonText</button>', $result);
     }
 
-    /**
-     * Test creating create microphone.
-     */
     public function test_create_microphone() {
         $result = create_microphone('testmic');
-        $this->assertEquals('<br></br><button id="record" class="record">Start</button><button id="stopRecord" class="stopRecord">Stop</button><button id="listenButton" class="listenButton">Listen recording</button>', $result); // phpcs:ignore moodle.Files.LineLength.MaxExceeded
+        $this->assertEquals('<br></br><button id="record" class="btn btn-primary record-btn">Start</button>'.
+            '<button id="stopRecord" class="btn btn-primary stopRecord-btn">Stop</button><button id="listenButton" '.
+            'class="btn btn-primary listen-btn">Listen recording</button>', $result);
     }
 
     /**
@@ -238,7 +295,7 @@ class locallib_test extends \advanced_testcase {
         global $USER;
 
         $context = \context_module::instance($this->digitala->cmid);
-        $assignment = new digitala_assignment($this->digitala->id, $context->id, $USER->id, $USER->username, 1, 1, $this->digitala->assignment, $this->digitala->resources, $this->digitala->attempttype, $this->digitala->attemptlang);
+        $assignment = new digitala_assignment($this->digitala->id, $context->id, $USER->id, $USER->username, 1, 1, $this->digitala->assignment, $this->digitala->resources, $this->digitala->attempttype, $this->digitala->attemptlang);  // phpcs:ignore moodle.Files.LineLength.MaxExceeded
 
         $fileinfo = array(
             'contextid' => context_user::instance($USER->id)->id,
@@ -267,7 +324,7 @@ class locallib_test extends \advanced_testcase {
         global $USER;
 
         $context = \context_module::instance($this->digitala->cmid);
-        $assignment = new digitala_assignment($this->digitala->id, $context->id, $USER->id, $USER->username, 4, 5, $this->digitala->assignment, $this->digitala->resources, $this->digitala->attempttype, $this->digitala->attemptlang);
+        $assignment = new digitala_assignment($this->digitala->id, $context->id, $USER->id, $USER->username, 4, 5, $this->digitala->assignment, $this->digitala->resources, $this->digitala->attempttype, $this->digitala->attemptlang);  // phpcs:ignore moodle.Files.LineLength.MaxExceeded
 
         $result = create_answerrecording_form($assignment);
         $this->assertMatchesRegularExpression('/form id="answerrecording"/', $result);
@@ -282,9 +339,9 @@ class locallib_test extends \advanced_testcase {
      */
     public function test_create_answerrecording_form_with_data() {
         global $USER;
-        answerrecording_form::mock_submit(array('audiostring' => '{"url":"http:\/\/localhost:8000\/draftfile.php\/5\/user\/draft\/0\/testing.wav","id": 0,"file":"testing.wav"}'), null, 'post', 'answerrecording_form');
+        answerrecording_form::mock_submit(array('audiostring' => '{"url":"http:\/\/localhost:8000\/draftfile.php\/5\/user\/draft\/0\/testing.wav","id": 0,"file":"testing.wav"}'), null, 'post', 'answerrecording_form');  // phpcs:ignore moodle.Files.LineLength.MaxExceeded
         $context = \context_module::instance($this->digitala->cmid);
-        $assignment = new digitala_assignment($this->digitala->id, $context->id, $USER->id, $USER->username, 1, 1, $this->digitala->assignment, $this->digitala->resources, $this->digitala->attempttype, $this->digitala->attemptlang);
+        $assignment = new digitala_assignment($this->digitala->id, $context->id, $USER->id, $USER->username, 1, 1, $this->digitala->assignment, $this->digitala->resources, $this->digitala->attempttype, $this->digitala->attemptlang);  // phpcs:ignore moodle.Files.LineLength.MaxExceeded
 
         $result = create_answerrecording_form($assignment);
         $this->assertEquals('No evaluation was found. Please return to previous page.', $result);
